[HUDI-9704] Move remaining APIs from reader context to record context
