+++ hudi-common/src/main/java/org/apache/hudi/common/engine/RecordContext.java
@@ -39,10 +39,12 @@ import org.apache.avro.generic.IndexedRecord;
 import javax.annotation.Nullable;
 
 import java.io.Serializable;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.function.BiFunction;
+import java.util.function.UnaryOperator;
 
 import static org.apache.hudi.common.model.HoodieRecord.HOODIE_IS_DELETED_FIELD;
 import static org.apache.hudi.common.model.HoodieRecord.RECORD_KEY_METADATA_FIELD;
@@ -272,6 +274,40 @@ public abstract class RecordContext<T> implements Serializable {
     Object deleteMarkerValue = getValue(record, deleteContext.getReaderSchema(), markerKeyValue.getLeft());
     return deleteMarkerValue != null
         && markerKeyValue.getRight().equals(deleteMarkerValue.toString());
+  }
+
+  /**
+   * Seals the engine-specific record to make sure the data referenced in memory do not change.
+   *
+   * @param record The record.
+   * @return The record containing the same data that do not change in memory over time.
+   */
+  public abstract T seal(T record);
+
+  /**
+   * Converts engine specific row into binary format.
+   *
+   * @param avroSchema The avro schema of the row
+   * @param record     The engine row
+   *
+   * @return row in binary format
+   */
+  public abstract T toBinaryRow(Schema avroSchema, T record);
+
+  /**
+   * Creates a function that will reorder records of schema "from" to schema of "to"
+   * all fields in "to" must be in "from", but not all fields in "from" must be in "to"
+   *
+   * @param from           the schema of records to be passed into UnaryOperator
+   * @param to             the schema of records produced by UnaryOperator
+   * @param renamedColumns map of renamed columns where the key is the new name from the query and
+   *                       the value is the old name that exists in the file
+   * @return a function that takes in a record and returns the record with reordered columns
+   */
+  public abstract UnaryOperator<T> projectRecord(Schema from, Schema to, Map<String, String> renamedColumns);
+
+  public final UnaryOperator<T> projectRecord(Schema from, Schema to) {
+    return projectRecord(from, to, Collections.emptyMap());
   }
 
   /**
