+++ hudi-common/src/main/java/org/apache/hudi/common/table/read/UpdateProcessor.java
@@ -89,7 +89,7 @@ public interface UpdateProcessor<T> {
           mergedRecord.setHoodieOperation(HoodieOperation.INSERT);
           readStats.incrementNumInserts();
         }
-        return mergedRecord.seal(readerContext);
+        return mergedRecord.seal(readerContext.getRecordContext());
       }
     }
   }
