+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/table/format/FlinkRowDataReaderContext.java
@@ -73,11 +73,6 @@ public class FlinkRowDataReaderContext extends HoodieReaderContext<RowData> {
   private final List<ExpressionPredicates.Predicate> predicates;
   private final Supplier<InternalSchemaManager> internalSchemaManager;
   private final HoodieTableConfig tableConfig;
-  // the converter is used to create a RowData contains primary key fields only
-  // for DELETE cases, it'll not be initialized if primary key semantics is lost.
-  // For e.g, if the pk fields are [a, b] but user only select a, then the pk
-  // semantics is lost.
-  private RecordKeyToRowDataConverter recordKeyRowConverter;
 
   public FlinkRowDataReaderContext(
       StorageConfiguration<?> storageConfiguration,
@@ -129,23 +124,6 @@ public class FlinkRowDataReaderContext extends HoodieReaderContext<RowData> {
   }
 
   @Override
-  public RowData seal(RowData rowData) {
-    if (rowData instanceof BinaryRowData) {
-      return ((BinaryRowData) rowData).copy();
-    }
-    return rowData;
-  }
-
-  @Override
-  public RowData toBinaryRow(Schema avroSchema, RowData record) {
-    if (record instanceof BinaryRowData) {
-      return record;
-    }
-    RowDataSerializer rowDataSerializer = RowDataAvroQueryContexts.getRowDataSerializer(avroSchema);
-    return rowDataSerializer.toBinaryRow(record);
-  }
-
-  @Override
   public ClosableIterator<RowData> mergeBootstrapReaders(
       ClosableIterator<RowData> skeletonFileIterator,
       Schema skeletonRequiredSchema,
@@ -191,25 +169,6 @@ public class FlinkRowDataReaderContext extends HoodieReaderContext<RowData> {
     };
   }
 
-  /**
-   * Creates a function that will reorder records of schema "from" to schema of "to".
-   * It's possible there exist fields in `to` schema, but not in `from` schema because of schema
-   * evolution.
-   *
-   * @param from           the schema of records to be passed into UnaryOperator
-   * @param to             the schema of records produced by UnaryOperator
-   * @param renamedColumns map of renamed columns where the key is the new name from the query and
-   *                       the value is the old name that exists in the file
-   * @return a function that takes in a record and returns the record with reordered columns
-   */
-  @Override
-  public UnaryOperator<RowData> projectRecord(Schema from, Schema to, Map<String, String> renamedColumns) {
-    RowType fromType = (RowType) RowDataAvroQueryContexts.fromAvroSchema(from).getRowType().getLogicalType();
-    RowType toType =  (RowType) RowDataAvroQueryContexts.fromAvroSchema(to).getRowType().getLogicalType();
-    RowProjection rowProjection = SchemaEvolvingRowDataProjection.instance(fromType, toType, renamedColumns);
-    return rowProjection::project;
-  }
-
   @Override
   public void setSchemaHandler(FileGroupReaderSchemaHandler<RowData> schemaHandler) {
     super.setSchemaHandler(schemaHandler);
@@ -229,7 +188,11 @@ public class FlinkRowDataReaderContext extends HoodieReaderContext<RowData> {
         .map(k -> Option.ofNullable(requiredSchema.getField(k)).map(Schema.Field::pos).orElse(-1))
         .mapToInt(Integer::intValue)
         .toArray();
-    recordKeyRowConverter = new RecordKeyToRowDataConverter(
+    // the converter is used to create a RowData contains primary key fields only
+    // for DELETE cases, it'll not be initialized if primary key semantics is lost.
+    // For e.g, if the pk fields are [a, b] but user only select a, then the pk
+    // semantics is lost.
+    RecordKeyToRowDataConverter recordKeyRowConverter = new RecordKeyToRowDataConverter(
         pkFieldsPos, (RowType) RowDataAvroQueryContexts.fromAvroSchema(requiredSchema).getRowType().getLogicalType());
     ((FlinkRecordContext) recordContext).setRecordKeyRowConverter(recordKeyRowConverter);
   }
